requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
pydantic>=2.0
pandas
matplotlib
python-dotenv
